samples: fix implicit int-to-float + CustomSTLAllocators warnings
* -Wimplicit-const-int-float-conversion: RAND_MAX (INT_MAX) was being converted
  to float in "rand()/RAND_MAX" expressions, losing precision. Cast RAND_MAX to
  FLOAT in the rnd() macros (BeginPush, SwapCallback) and at the inline division
  sites (BackBufferScale, BillBoard, Explosion). Explosion's integer switch on
  rand()*4/RAND_MAX is deliberately left as integer division.

* CustomSTLAllocators: add 'typename' to the dependent allocator types
  (Alloc::pointer/const_pointer/size_type/rebind<>::other) (-Wtypename-missing);
  rename the converting-ctor template parameter so it no longer shadows the
  class template's Alignment (-Wmicrosoft-template-shadow). The deprecated
  classic-allocator members (address/allocate/construct/destroy/max_size/rebind)
  are exactly what this sample demonstrates, so -Wdeprecated-declarations is
  silenced file-locally with a rationale rather than gutting the sample.

+++ XDKSamples/Graphics/BackBufferScale/BackBufferScale.cpp
@@ -48,7 +48,7 @@ struct Tree
 
 inline FLOAT GaussianRand( FLOAT min, FLOAT max, FLOAT width )
 {
-    FLOAT x = (FLOAT)(rand()-rand())/RAND_MAX;
+    FLOAT x = (FLOAT)(rand()-rand())/(FLOAT)RAND_MAX;
     return (max-min)*expf( -width*x*x) + min;
 }
 
@@ -137,14 +137,14 @@ CXBoxSample::CXBoxSample()
     for( DWORD i=0; i<NUM_TREES; i++ )
     {
         // Position and size the trees randomly
-        FLOAT z = 5.0f + ( 60.0f * rand() ) / RAND_MAX;
-        FLOAT x = ( 0.6f * z * (rand()-rand()) ) / RAND_MAX;
+        FLOAT z = 5.0f + ( 60.0f * rand() ) / (FLOAT)RAND_MAX;
+        FLOAT x = ( 0.6f * z * (rand()-rand()) ) / (FLOAT)RAND_MAX;
         FLOAT w = GaussianRand( 5.0f, 2.0f, 2.0f );
         FLOAT h = GaussianRand( 7.0f, 4.0f, 2.0f );
 
         // Each tree is a random color between red and green
-        DWORD r = (DWORD)(255 * ( (0.25f) + (0.75f*rand())/RAND_MAX) );
-        DWORD g = (DWORD)(255 * ( (0.25f) + (0.75f*rand())/RAND_MAX) );
+        DWORD r = (DWORD)(255 * ( (0.25f) + (0.75f*rand())/(FLOAT)RAND_MAX) );
+        DWORD g = (DWORD)(255 * ( (0.25f) + (0.75f*rand())/(FLOAT)RAND_MAX) );
         DWORD b = 0x00;
         DWORD color = (0xff<<24)|(r<<16)|(g<<8)|(b);
